// ui/src/client/types.gen.ts
export type ClientContactCreate = {
    /**
     * Name
     */
    name: string;
    /**
     * Email
     */
    email?: string | null;
    /**
     * Phone
     */
    phone?: string | null;
    /**
     * Role Title
     */
    role_title?: string | null;
};
// ...
export type ClientCreate = {
    /**
     * Name
     */
    name: string;
    /**
     * Address
     */
    address?: string | null;
    /**
     * Contacts
     */
    contacts?: Array<ClientContactCreate>;
};